requirements.txt
flask==2.3.3
flask-cors==4.0.0
psycopg2-binary==2.9.11
cryptography==41.0.4
werkzeug==2.3.7
python-dotenv==1.0.1
